# speed up `group_by(replicate)`